ai-image-forensics-platform/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1
opencv-python-headless==4.9.0.80
Pillow==10.2.0
scipy==1.12.0
numpy==1.26.4
redis==5.0.3
celery==5.3.6
prometheus-client==0.20.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
python-magic==0.4.27
